Made Gson type adapters better extensible
